requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.6
django-allauth==0.38.0
django-cors-headers==2.4.0
django-heroku==0.3.1
django-rest-auth==0.9.3
django-taggit==0.23.0
django-webpack-loader==0.6.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.0.1
Pillow==5.4.1
psycopg2==2.7.6.1
pycparser==2.19
PyJWT==1.7.1
pylint==2.2.2
pyOpenSSL==19.0.0
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.11.0
